+++ compiler/src/main/kotlin/dev/zacsweers/metro/compiler/ir/transformers/MembersInjectorTransformer.kt
@@ -26,6 +26,7 @@ import dev.zacsweers.metro.compiler.ir.rawTypeOrNull
 import dev.zacsweers.metro.compiler.ir.regularParameters
 import dev.zacsweers.metro.compiler.ir.requireSimpleFunction
 import dev.zacsweers.metro.compiler.ir.thisReceiverOrFail
+import dev.zacsweers.metro.compiler.ir.trackFunctionCall
 import org.jetbrains.kotlin.ir.builders.IrBlockBodyBuilder
 import org.jetbrains.kotlin.ir.builders.irBlockBody
 import org.jetbrains.kotlin.ir.builders.irGet
@@ -236,6 +237,7 @@ internal class MembersInjectorTransformer(context: IrMetroContext) : IrMetroCont
         pluginContext.createIrBuilder(symbol).irBlockBody {
           addMemberInjection(
             context = metroContext,
+            callingFunction = this@apply,
             instanceReceiver = regularParameters[0],
             injectorReceiver = dispatchReceiverParameter!!,
             injectFunctions = injectFunctions,
@@ -254,12 +256,14 @@ internal class MembersInjectorTransformer(context: IrMetroContext) : IrMetroCont
 
 internal fun IrBlockBodyBuilder.addMemberInjection(
   context: IrMetroContext,
+  callingFunction: IrSimpleFunction,
   injectFunctions: Map<IrSimpleFunction, Parameters<MembersInjectParameter>>,
   parametersToFields: Map<Parameter, IrField>,
   instanceReceiver: IrValueParameter,
   injectorReceiver: IrValueParameter,
 ) {
   for ((function, parameters) in injectFunctions) {
+    context.trackFunctionCall(callingFunction, function)
     +irInvoke(
       dispatchReceiver = irGetObject(function.parentAsClass.symbol),
       callee = function.symbol,